+++ src/test/java/calculator_test/TangTest.java
@@ -7,7 +7,7 @@ import org.testng.annotations.Test;
 public class TangTest extends BaseTest {
 
     @Test(dataProvider = "valuesForTangTest", groups = {"unitTests1"})
-    public void tangTest(long a, double expectedValue) {
+    public void tangTest(double a, double expectedValue) {
         double result = calculator.tg(a);
         Assert.assertEquals(result, expectedValue, "Invalid tang value");
     }
